Mostly Bug Fixes
- Reworked updating the Calm Lands Bitmask to stop bugs seen with getting Caladbolg and Nirvana Chest
- Change SpawnPoint from byte to short and updated values accordingly
- Fixed Wantz not appearing in Macalania bug
- Added YojimboFaythTransition but currently not active as it doesn't work right now
- Added a fix for a really old bug where the Biran and Yenke encounter could turn off no encounters until the armor is re-equipped

+++ FFXCutsceneRemover/Components/PreviousGameState.cs
@@ -13,7 +13,7 @@ namespace FFXCutsceneRemover
          * If we only ever write the value then there is no need to add it here. */
         public short? RoomNumber = null;
         public short? Storyline = null;
-        public short? SpawnPoint = null;
+        public byte? SpawnPoint = null;
         public short? BattleState = null;
         public short? BattleState2 = null;
         public short? Input = null;
